Fix repeated initial characters in annotation.
Turns out using a non-capture group is not a perfect replacement for
simulating negative lookbehind. Even though it's not capture, it counts
in calculating the length & index of the match. We have to explicitly
account for the length of the "don't match here" group to distinguish
matches at the beginning and one character in to a text segment.