Youtube iframe fix
Now the youtube player status is inside event.data and the player is
inside event.target.
I added the videoIsMuted variable in order to keep track of the muted
state of the player.
It seems that the changes made by the player.mute() and player.unMute()
functions aren't immediately available to the player.isMuted() function.
I couldn't get it to recognize that the muted state of the player had
changed within that call stack. Therefore, the audio would mute, but the
ayamel media player would receive false from the YouTubePlayer.muted
getter function. Then it would update after it exited that call stack
and the next time you press mute, the player.isMuted() function would
return true and the ayamel player would show the mute button as
currently muted.